kuberlnetes (kuberlnetes v0.1.0)

kuberlnetes A low level kubernetes api helper - primarlily for applications that interact with the kubernetes api from inside the cluster.

Types

http_status_err/0

-type http_status_err() :: 300..600.

kube_cfg_options/0

-type kube_cfg_options() :: #{path => string(), context => string(), cache => boolean()}.

maybe/1

-type maybe(T) :: T | undefined.

mutation_request/0

-type mutation_request() :: #{path => string(), body => map() | binary()}.

options/0

-type options() :: #{server => server() | undefined}.

resp_with_body/0

-type resp_with_body() :: {ok, map()} | {error, http_status_err()}.

server/0

-type server() :: #server{}.

Functions

delete(Path, Opts)

-spec delete(Path, Opts) -> {ok, Body} when Path :: string(), Opts :: options(), Body :: map().

Deletes the given resource

from_config()

from_config(Options)

-spec from_config(Options) -> server() when Options :: kube_cfg_options().

Loads the current-context from ~/.kube/config

from_raw(Url)

-spec from_raw(string()) -> server().

get(Path)

-spec get(Path) -> {ok, Body} when Path :: string(), Body :: map().

get(Path, Opts)

-spec get(Path, Opts) -> {ok, Body} when Path :: string(), Opts :: options(), Body :: map().

Initiates a GET request against the configured kubernetes api server or raises an error if the server is not configured correctly.

headers(Server)

headers(Server, AdditionalHeaders)

in_cluster()

-spec in_cluster() -> server().

Configures a server using the service account

microtime_now()

-spec microtime_now() -> binary().

Returns the current datetime in the kubernetes MicroTime format see: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.22/#microtime-v1-meta

patch(Request, Opts)

-spec patch(Request, Opts) -> {ok, Body}
         when Request :: mutation_request(), Opts :: options(), Body :: map().

Initiates a PATCH request against the configured kubernetes api server or raises an error if the server is not configured correctly. see: server()

post(Request, Opts)

-spec post(Request, Opts) -> {ok, Body}
        when Request :: mutation_request(), Opts :: options(), Body :: map().

Initiates a POST request against the configured kubernetes api server or raises an error if the server is not configured correctly. see: server()

watch(Opts)

-spec watch(Opts) -> Result when Opts :: kuberlnetes_watch:watch_opts(), Result :: {ok, pid()}.

watch_close(Pid)

-spec watch_close(Pid) -> Result when Pid :: pid(), Result :: ok.

Closes an active watch process when not running in a supervision tree.

watches()

-spec watches() -> Result when Result :: [pid()].

Returns all currently active watch processes